requirements.txt
Flask
torch
transformers
google-generativeai